moving ShiftSummaryData to separate class
