+++ client/src/main/scala/drt/client/services/handlers/ShiftSummaryStaffingHandler.scala
@@ -3,7 +3,6 @@ package drt.client.services.handlers
 import diode.AnyAction.aType
 import diode.{ActionResult, Effect, ModelRW, NoAction}
 import diode.data.{Pot, Ready}
-import drt.client.components
 import drt.client.logger.log
 import drt.client.services.DrtApi
 import drt.shared.ShiftAssignments
@@ -26,50 +25,6 @@ case class UpdateShiftSummaryStaffingWithAssignment(shiftAssignments: ShiftAssig
 
 case class SetShiftSummaryStaffing(staffShifts: Seq[ShiftSummaryData.ShiftSummaryStaffing])
 
-object ShiftSummaryData {
-  case class ShiftDate(
-                        year: Int,
-                        month: Int,
-                        day: Int,
-                        hour: Int,
-                        minute: Int
-                      ) {
-    def toClientShiftDate: components.ShiftDate = components.ShiftDate(year, month, day, hour, minute)
-  }
-
-  case class StaffTableEntry(
-                              column: Int,
-                              row: Int,
-                              name: String,
-                              staffNumber: Int,
-                              startTime: ShiftDate,
-                              endTime: ShiftDate) {
-    def toClientStaffTableEntry: components.StaffTableEntry = {
-      components.StaffTableEntry(column, row, name, staffNumber, startTime.toClientShiftDate, endTime.toClientShiftDate)
-    }
-  }
-
-  case class ShiftSummary(
-                           name: String,
-                           defaultStaffNumber: Int,
-                           startTime: String,
-                           endTime: String
-                         ) {
-    def toClientShiftSummary: components.ShiftSummary =
-      components.ShiftSummary(name, defaultStaffNumber, startTime, endTime)
-
-  }
-
-  case class ShiftSummaryStaffing(
-                                   index: Int,
-                                   shiftSummary: ShiftSummary,
-                                   staffTableEntries: Seq[StaffTableEntry]
-                                 ) {
-    def toClientShiftSummaryStaffing: components.ShiftSummaryStaffing =
-      components.ShiftSummaryStaffing(index, shiftSummary.toClientShiftSummary, staffTableEntries.map(_.toClientStaffTableEntry))
-  }
-
-}
 
 class ShiftSummaryStaffingHandler[M](modelRW: ModelRW[M, Pot[Seq[ShiftSummaryData.ShiftSummaryStaffing]]]) extends LoggingActionHandler(modelRW) {
